requirements.txt
alabaster==0.7.13
alembic==1.9.4
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
fastapi==0.89.1
filelock==3.9.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpie==3.2.1
httptools==0.5.0
httpx==0.23.3
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==0.4.3
orjson==3.8.5
packaging==23.0
pathspec==0.11.0
platformdirs==2.6.2
pluggy==1.0.0
prompt-toolkit==3.0.38
py==1.11.0
pydantic==1.10.4
Pygments==2.14.0
PySocks==1.7.1
pytest==7.1.3
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==13.3.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==2.0.4
sqlparse==0.4.3
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
vine==5.0.0
virtualenv==20.20.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
